31 user related classes repo service controller

+++ src/main/java/be/addergebroed/weeklymealplanner/recipe/model/RecipeIngredient.java
@@ -19,15 +19,15 @@ public class RecipeIngredient {
     private Long Id;
 
     @Column(nullable = false)
-    private int qty;
+    private double qty;
 
     @ManyToOne
-    @JoinColumn(name="unit_id", nullable = false)
+    @JoinColumn(name = "unit_id", nullable = false)
     private FoodUnit unit;
 
-@ManyToOne
-    @JoinColumn(name="ingredient_id",nullable = false)
-    private    Ingredient ingredient;
+    @ManyToOne
+    @JoinColumn(name = "ingredient_id", nullable = false)
+    private Ingredient ingredient;
 
     @ManyToMany(mappedBy = "recipeIngredients")
     private Set<Recipe> recipes;
